refactor more and fix linter
